Start split of SPI code
