+++ java/execute/src/com/ibm/streamsx/topology/internal/spljava/Schemas.java
@@ -13,7 +13,7 @@ import com.ibm.streams.operator.StreamSchema;
 import com.ibm.streams.operator.types.Blob;
 import com.ibm.streams.operator.types.XML;
 import com.ibm.streamsx.topology.internal.core.ObjectSchemas;
-import com.ibm.streamsx.topology.spi.TupleSerializer;
+import com.ibm.streamsx.topology.spi.runtime.TupleSerializer;
 
 public class Schemas {
     @SuppressWarnings("unused")
