+++ java/execute/src/com/ibm/streamsx/topology/spi/operators/Source.java
@@ -7,8 +7,8 @@ import com.ibm.streams.operator.model.SharedLoader;
 import com.ibm.streamsx.topology.function.FunctionContext;
 import com.ibm.streamsx.topology.function.Supplier;
 import com.ibm.streamsx.topology.internal.functional.ops.FunctionSource;
-import com.ibm.streamsx.topology.spi.FunctionalOperator;
 import com.ibm.streamsx.topology.spi.Invoker;
+import com.ibm.streamsx.topology.spi.runtime.FunctionalOperator;
 
 /**
  * A functional source operator.
